% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_table_MSWord.R
\name{export_table_MSWord}
\alias{export_table_MSWord}
\alias{export_table_MSWord.list}
\alias{export_table_MSWord.test_mediation}
\title{Export result table to Word}
\usage{
export_table_MSWord(test_model, ...)

\method{export_table_MSWord}{list}(test_model, orientation = c("landscape", "portrait"), ...)

\method{export_table_MSWord}{test_mediation}(test_model, digits = 4, ...)
}
\arguments{
\item{test_model}{an object inheriting from class
\code{"\link{test_mediation}"} or a list of objects of that class.
of object}

\item{orientation}{the  orientation in which the table is displayed in the
document. When set to portrait, the tables are displayed
underneath each other. When set to landscape, the tables are displayed side
by side.}

\item{digits}{a positive integer, which determines the number of decimals
that should be displayed in the table. The default is to display 4 decimals.}
}
\value{
An object of class \code{"\link{rdocx}"}, containing a table of the
results of the provided objects.
}
\description{
Export the table containing the results of mediation analysis to a Microsoft Word
document.
}
\examples{
data("BSG2014")

boot_robust <- test_mediation(TeamCommitment ~
                                m(TaskConflict) +
                                  ValueDiversity,
                              data = BSG2014)

table <- export_table_MSWord(boot_robust)
print(table, target = "table_robust.docx")


boot_ols <- test_mediation(TeamCommitment ~
                                m(TaskConflict) +
                                  ValueDiversity,
                              data = BSG2014,
                              robust = FALSE)

tables <- export_table_MSWord(list(boot_robust, boot_ols))

print(tables, target = "tables_landscape.docx")


}
\author{
Vincent Drenth
}
