% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagram.R
\name{model_diagram}
\alias{model_diagram}
\alias{model_diagram.fit_mediation}
\alias{model_diagram.test_mediation}
\alias{model_diagram.default}
\alias{model_diagram.setup_model_diagram}
\title{Diagram of a mediation model}
\usage{
model_diagram(object, ...)

\method{model_diagram}{fit_mediation}(object, width = 8, height = 2, ...)

\method{model_diagram}{test_mediation}(object, width = 8, height = 2, ...)

\method{model_diagram}{default}(
  object,
  x = object,
  y,
  m,
  covariates = NULL,
  model = c("parallel", "serial"),
  width = 8,
  height = 2,
  ...
)

\method{model_diagram}{setup_model_diagram}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class
\code{"\link[robmed]{fit_mediation}"} or
\code{"\link[robmed]{test_mediation}"} containing results from (robust)
mediation analysis.  For the default method, the names of the independent
variables of interest can also be specified via this argument.}

\item{\dots}{additional arguments to be passed down, eventually to
\code{\link[ggplot2]{geom_rect}()}, \code{\link[ggplot2]{geom_text}()},
and \code{\link[ggplot2]{geom_segment}()}.}

\item{width, height}{numeric values giving the width and height of the boxes
representing the variables in the model diagram.}

\item{x}{a character vector specifying the names of the independent
variables of interest to be displayed in the model diagram.}

\item{y}{a character string specifying the name of the dependent
variable to be displayed in the model diagram.}

\item{m}{a character vector specifying the names of the hypothesized
mediator variables to be displayed in the model diagram.}

\item{covariates}{a character vector specifying the names of additional
covariates to be included as control variables in the model diagram.}

\item{model}{a character string specifying the type of model in case of
multiple mediators.  Possible values are \code{"parallel"} (the default) for
the parallel multiple mediator model, or \code{"serial"} for the serial
multiple mediator model.  This is only relevant for drawing a diagram of a
model with multiple hypothesized mediators.}
}
\value{
An object of class \code{"\link[ggplot2]{ggplot}"}.
}
\description{
Draw a diagram of a mediation model. Control variables are by default drawn
in gray to visually distinguish them from the main variables of interest.
}
\details{
Methods first call \code{\link{setup_model_diagram}()} to obtain all
necessary information to produce the plot, then the
\code{"setup_model_diagram"} method is called to produce the plot.
}
\examples{
## drawing diagrams from scratch

# simple mediation model
model_diagram(x = "ValueDiversity",
              y = "TeamCommitment",
              m = "TaskConflict",
              size = 3.5)

# serial multiple mediators
model_diagram(x = "ValueDiversity",
              y = "TeamScore",
              m = c("TaskConflict", "TeamCommitment"),
              model = "serial",
              size = 3.5)

# parallel multiple mediators and control variables
model_diagram(x = "SharedLeadership",
              y = "TeamPerformance",
              m = c("ProceduralJustice", "InteractionalJustice"),
              covariates = c("AgeDiversity", "GenderDiversity"),
              model = "parallel",
              size = 3.5)


## draw a diagram from an object containing results from robust
## mediation analysis

# load data
data("BSG2014")
# seed to be used for the random number generator
seed <- 20211117
# perform mediation analysis via robust bootstrap test ROBMED
set.seed(seed)
robust_boot <- test_mediation(BSG2014,
                              x = "ValueDiversity",
                              y = "TeamCommitment",
                              m = "TaskConflict",
                              robust = TRUE)
# create model diagram
model_diagram(robust_boot, size = 3.5)

}
\seealso{
\code{\link{fit_mediation}()}, \code{\link{test_mediation}()},
\code{\link{setup_model_diagram}()}
}
\author{
Andreas Alfons
}
