\name{robmedShiny-package}
\alias{robmedShiny-package}
\docType{package}
\title{
\packageTitle{robmedShiny}
}
\description{
\packageDescription{robmedShiny}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{robmedShiny}
\packageIndices{robmedShiny}
}
\examples{
## compare bootstrap methods for mediation analysis on simulated data
\dontrun{
run_shiny_app()
}
}
\references{
Alfons, A., Ates, N.Y. and Groenen, P.J.F. (2021) A robust bootstrap test
for mediation analysis.  \emph{Organizational Research Methods},
\doi{10.1177/1094428121999096}.
}
\author{
\packageAuthor{robmedShiny}

Maintainer: \packageMaintainer{robmedShiny}
}
\keyword{package}
