% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_model_diagram.R
\name{setup_model_diagram}
\alias{setup_model_diagram}
\alias{setup_model_diagram.fit_mediation}
\alias{setup_model_diagram.test_mediation}
\alias{setup_model_diagram.default}
\title{Set up information for a diagram of a mediation model}
\usage{
setup_model_diagram(object, ...)

\method{setup_model_diagram}{fit_mediation}(object, ...)

\method{setup_model_diagram}{test_mediation}(object, ...)

\method{setup_model_diagram}{default}(
  object,
  x = object,
  y,
  m,
  covariates = NULL,
  model = c("parallel", "serial"),
  width = 8,
  height = 2,
  ...
)
}
\arguments{
\item{object}{an object inheriting from class
\code{"\link[robmed]{fit_mediation}"} or
\code{"\link[robmed]{test_mediation}"} containing results from (robust)
mediation analysis.  For the default method, the names of the independent
variables of interest can also be specified via this argument.}

\item{\dots}{additional arguments to be passed down.}

\item{x}{a character vector specifying the names of the independent
variables of interest to be displayed in the model diagram.}

\item{y}{a character string specifying the name of the dependent
variable to be displayed in the model diagram.}

\item{m}{a character vector specifying the names of the hypothesized
mediator variables to be displayed in the model diagram.}

\item{covariates}{a character vector specifying the names of additional
covariates to be included as control variables in the model diagram.}

\item{model}{a character string specifying the type of model in case of
multiple mediators.  Possible values are \code{"parallel"} (the default) for
the parallel multiple mediator model, or \code{"serial"} for the serial
multiple mediator model.  This is only relevant for setting up a diagram of
a model with multiple hypothesized mediators.}

\item{width, height}{numeric values giving the width and height of the boxes
representing the variables in the model diagram.}
}
\value{
An object of class \code{"setup_model_diagram"} with the following
components:
\item{boxes}{a data frame containing the coordinates and other relevant
information on the boxes representing the variables in the model diagram.}
\item{arrows}{a data frame containing the coordinates and other relevant
information on the arrows representing the connections between the variables
in the model diagram.}
\item{colors}{a character vector of length two giving the colors to be used
for main variables of interest and for control variables, respectively.}
\item{size}{a numeric value giving the default text size.}
\item{expand}{a numeric vector giving the default scale expansion vectors
for the plot, see \code{\link[ggplot2]{expansion}()}.}
}
\description{
Construct the relevant information for drawing a diagram of a mediation
model.
}
\details{
This function is used internally by \code{\link{model_diagram}()}.  It
may also be useful for users who want to produce a similar plot, but who
want more control over what information to display or how to display that
information.

Specifically, a user can modify the positions of the boxes representing the
variables in the \code{boxes} component of the returned object.  However,
note that the arrows connecting the boxes in the \code{arrows} component
need to be modified accordingly.
}
\examples{
## parallel multiple mediators and control variables

# set up information for the model diagram
setup <- setup_model_diagram(x = "ValueDiversity",
                             y = "TeamCommitment",
                             m = "TaskConflict")

# draw the diagram
model_diagram(setup)

# draw a similar diagram, but without the boxes
arrow_head <- grid::arrow(angle = 15, length = unit(8, "pt"),
                          ends = "last", type = "closed")
ggplot() +
  geom_segment(mapping = aes(x = x, y = y, xend = xend,
                             yend = yend),
               data = setup$arrows, arrow = arrow_head) +
  geom_text(mapping = aes(x = x, y = y, label = Label),
            data = setup$boxes) +
  coord_fixed() +
  xlim(min(setup$boxes$xmin), max(setup$boxes$xmax)) +
  ylim(min(setup$boxes$ymin), max(setup$boxes$ymax)) +
  theme_void()

}
\seealso{
\code{\link{fit_mediation}()}, \code{\link{test_mediation}()},
\code{\link{model_diagram}()}
}
\author{
Andreas Alfons
}
