% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_flextable.R
\name{to_flextable}
\alias{to_flextable}
\title{Tabular summary of results from (robust) mediation analysis}
\usage{
to_flextable(object, ...)
}
\arguments{
\item{object}{an object inheriting from class
\code{"\link[robmed]{test_mediation}"} or
\code{"\link[robmed:summary.test_mediation]{summary_test_mediation}"}
containing results from (robust) mediation analysis, or a list of such
objects.}

\item{dots}{additional arguments to be passed down.}
}
\description{
Convert results from (robust) mediation analysis to a tabular summary in
the form of a \code{\link[flextable]{flextable}}.  This \code{flextable}
can easily be exported to Microsoft Word or integrated into a dynamic
\proglang{Markdown} document, which eliminates the risk of mistakes in
reporting that stem from copying-and-pasting results.
}
