% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_latex.R
\name{to_latex}
\alias{to_latex}
\title{Generate LaTeX code for table}
\usage{
to_latex(test_model, digits = 4)
}
\arguments{
\item{test_model}{an object inheriting from class
\code{"\link{test_mediation}"} or a list of objects of that class.}

\item{digits}{a positive integer, which determines the number of decimals
that should be displayed in the table. The default is to display 4 decimals.}
}
\value{
A character object containing the latex code that produces a table of
the results of the test_mediation object.
}
\description{
Generates LaTeX code that displays a table containing results of mediation
analyis.
}
