\name{robmedShiny-package}
\alias{robmedShiny-package}
\docType{package}
\title{
\packageTitle{robmedShiny}
}
\description{
\packageDescription{robmedShiny}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{robmedShiny}
\packageIndices{robmedShiny}
}
\examples{
\dontrun{

# start graphical user interface
start_GUI()

# run shiny app for simulating data
run_shiny_app("simulation")
}
}
\references{
Alfons, A., Ates, N.Y. and Groenen, P.J.F. (2022a) A Robust Bootstrap Test for
Mediation Analysis.  \emph{Organizational Research Methods}, \bold{25}(3),
591--617.  doi:10.1177/1094428121999096.

Alfons, A., Ates, N.Y. and Groenen, P.J.F. (2022b) Robust Mediation Analysis:
The \R Package \pkg{robmed}.  \emph{Journal of Statistical Software},
\bold{103}(13), 1--45.  doi:10.18637/jss.v103.i13.
}
\author{
\packageAuthor{robmedShiny}

Maintainer: \packageMaintainer{robmedShiny}
}
\keyword{package}
