% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\alias{start_GUI}
\title{Shiny apps: Graphical user interface and data simulation}
\usage{
run_shiny_app(which = c("GUI", "simulation"))

start_GUI()
}
\arguments{
\item{which}{a character string indicating which \pkg{shiny} app to run.
Possible values are \code{"GUI"} (the default) to start a graphical user
interface for package \pkg{robmed}, or \code{"simulation"} to run an app
for exploring the behavior of various bootstrap procedures for mediation
analysis on simulated data.}
}
\description{
Run one of the following \pkg{shiny} apps: a graphical user interface
for package \pkg{robmed}, or an app for exploring the behavior of various
bootstrap procedures for mediation analysis on simulated data.
}
\details{
Function \code{start_GUI()} is a simple wrapper for
\code{run_shiny_app(which = "GUI")} to start the graphical user interface
for package \pkg{robmed}.

This graphical user interface allows to select or import a data set, and to
perform the robust bootstrap test ROBMED (Alfons et al., 2022a) or the OLS
bootstrap test (e.g., Preacher & Hayes, 2004, 2008).

In the app for exploring the behavior of  various bootstrap procedures for
mediation analysis on simulated data (\code{which = "simulation"}), the
default settings follow the simulation design of Zu & Yuan (2010).  You can
adjust the total number of observations, the values of the coefficients in
the mediation model, the number of outliers, as well as the outlier shifts
in the individual variables.

For each selected methods, the bootstrap distribution of the indirect
effect is shown together with a shaded area representing the 95\% confidence
interval.  If the app does not update quickly enough, try to reduce the
number of bootstrap samples.
}
\examples{
\dontrun{

# start graphical user interface
start_GUI()

# run shiny app for simulating data
run_shiny_app("simulation")
}

}
\references{
Alfons, A., Ates, N.Y. and Groenen, P.J.F. (2022a) A Robust Bootstrap Test
for Mediation Analysis.  \emph{Organizational Research Methods},
\bold{25}(3), 591--617.  doi:10.1177/1094428121999096.

Alfons, A., Ates, N.Y. and Groenen, P.J.F. (2022b) Robust Mediation
Analysis: The \R Package \pkg{robmed}.  \emph{Journal of Statistical
Software}, \bold{103}(13), 1--45.  doi:10.18637/jss.v103.i13.

Preacher, K.J. and Hayes, A.F. (2004) SPSS and SAS Procedures for Estimating
Indirect Effects in Simple Mediation Models.  \emph{Behavior Research
Methods, Instruments, & Computers}, \bold{36}(4), 717--731.
doi:10.3758/bf03206553.

Preacher, K.J. and Hayes, A.F. (2008) Asymptotic and Resampling Strategies
for Assessing and Comparing Indirect Effects in Multiple Mediator Models.
\emph{Behavior Research Methods}, \bold{40}(3), 879--891.
doi:10.3758/brm.40.3.879.

Zu, J. and Yuan, K.-H. (2010) Local Influence and Robust Procedures for
Mediation Analysis.  \emph{Multivariate Behavioral Research}, \bold{45}(1),
1--44.  doi:10.1080/00273170903504695.
}
\seealso{
\code{\link{test_mediation}}
}
\author{
Andreas Alfons
}
\keyword{documentation}
