% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_flextable.R
\name{to_flextable}
\alias{to_flextable}
\title{Create flextable from mediation test}
\usage{
to_flextable(test_model, ...)
}
\arguments{
\item{test_model}{an object inheriting from class
\code{"\link{test_mediation}"} or a list of objects of that class.
of object}

\item{digits}{a positive integer, which determines the number of decimals
that should be displayed in the table. The default is to display 4 decimals.}

\item{merged}{boolean that determines whether the flextables are merged to show
the results of two tests in one table (except possibly the last one) or
seperate. Only used when test_model
is a list.}

\item{p_values}{boolean that indicates whether the p-values for indirect
effects should be included in the flextable or not.
Default is to include the p-values}
}
\value{
An object of class \code{"\link{flextable}"} or a list of objects of
this class.
}
\description{
Creates a flextable from a test_mediation object
}
\examples{
data("BSG2014")

boot_robust <- test_mediation(TeamCommitment ~
                                m(TaskConflict) +
                                  ValueDiversity,
                              data = BSG2014)

ft <- to_flextable(boot_robust)

boot_ols <- test_mediation(TeamCommitment ~
                                m(TaskConflict) +
                                  ValueDiversity,
                              data = BSG2014,
                              robust = FALSE)

flextables_list <- to_flextable(list(boot_robust, boot_ols))


}
\author{
Vincent Drenth
}
